_archive/data_api/data_api/api_query_engine.py
import typing
import strawberry


@strawberry.type
class Book:
    title: str
    author: str


@strawberry.type
class Query:
    books: typing.List[Book]


def get_books():
    return [
        Book(title='The Great Gatsby', author='F. Scott Fitzgerald',),
    ]


@strawberry.type
class Query:
    books: typing.List[Book] = strawberry.field(resolver=get_books)


schema = strawberry.Schema(query=Query)
